Move updatePayments to HomeScreen and refresh payments  after updating
